Fix stan

+++ lib/term/src/EventParser.php
@@ -2,7 +2,6 @@
 
 namespace PhpTui\Term;
 
-use ParseError;
 use PhpTui\Term\Event\CharKeyEvent;
 use PhpTui\Term\Event\FocusEvent;
 use PhpTui\Term\Event\FunctionKeyEvent;
@@ -237,7 +236,7 @@ final class EventParser
      */
     private function parseCsiModifierKeyCode(array $buffer): Event
     {
-        $str = implode('', array_slice($buffer, 2, array_key_last($buffer)));
+        $str = implode('', array_slice($buffer, 2, (int)array_key_last($buffer)));
         // split string into bytes
         $parts = (array)explode(';', $str);
